django_backend/pyproject.toml
[project]
name = "django_backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Xiron",email = "xiron7077@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "mpi4py (>=4.0.3,<5.0.0)",
    "cupy-cuda12x (>=13.4.1,<14.0.0)",
    "numba (>=0.61.2,<0.62.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
